v3_agentic_rag/requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-chroma==0.1.4
langgraph==0.2.20
chromadb==0.5.3
tavily-python==0.5.0
rank_bm25==0.2.2
streamlit==1.39.0
python-dotenv==1.0.1
pypdf==5.0.0